+++ src/components/Parameters/Parameters.tsx
@@ -67,7 +67,10 @@ function DropdownWithinHeader(props) {
   );
 }
 
-export function BodyContent(props: { content: MediaContentModel; description?: string }): JSX.Element {
+export function BodyContent(props: {
+  content: MediaContentModel;
+  description?: string;
+}): JSX.Element {
   const { content, description } = props;
   const { isRequestType } = content;
   return (
@@ -76,7 +79,12 @@ export function BodyContent(props: { content: MediaContentModel; description?: s
         return (
           <>
             {description !== undefined && <Markdown source={description} />}
-            <Schema skipReadOnly={isRequestType} key="schema" schema={schema} />
+            <Schema
+              skipReadOnly={isRequestType}
+              skipWriteOnly={!isRequestType}
+              key="schema"
+              schema={schema}
+            />
           </>
         );
       }}