+++ src/services/models/Operation.ts
@@ -25,6 +25,7 @@ import { FieldModel } from './Field';
 import { MediaContentModel } from './MediaContent';
 import { RequestBodyModel } from './RequestBody';
 import { ResponseModel } from './Response';
+import { SideNavStyleEnum } from '../RedocNormalizedOptions';
 
 export interface XPayloadSample {
   lang: 'payload';
@@ -49,6 +50,7 @@ export class OperationModel implements IMenuItem {
   id: string;
   absoluteIdx?: number;
   name: string;
+  sidebarLabel: string;
   description?: string;
   type = 'operation' as const;
 
@@ -104,11 +106,13 @@ export class OperationModel implements IMenuItem {
 
     this.name = getOperationSummary(operationSpec);
 
+    this.sidebarLabel = options.sideNavStyle === SideNavStyleEnum.PathOnly ? this.path : this.name;
+
     if (this.isCallback) {
       // NOTE: Callbacks by default should not inherit the specification's global `security` definition.
       // Can be defined individually per-callback in the specification. Defaults to none.
       this.security = (operationSpec.security || []).map(
-        (security) => new SecurityRequirementModel(security, parser),
+        security => new SecurityRequirementModel(security, parser),
       );
 
       // TODO: update getting pathInfo for overriding servers on path level
@@ -122,7 +126,7 @@ export class OperationModel implements IMenuItem {
           : this.pointer;
 
       this.security = (operationSpec.security || parser.spec.security || []).map(
-        (security) => new SecurityRequirementModel(security, parser),
+        security => new SecurityRequirementModel(security, parser),
       );
 
       this.servers = normalizeServers(
@@ -173,7 +177,8 @@ export class OperationModel implements IMenuItem {
   @memoize
   get requestBody() {
     return (
-      this.operationSpec.requestBody && new RequestBodyModel({
+      this.operationSpec.requestBody &&
+      new RequestBodyModel({
         parser: this.parser,
         infoOrRef: this.operationSpec.requestBody,
         options: this.options,
@@ -218,7 +223,7 @@ export class OperationModel implements IMenuItem {
       this.operationSpec.pathParameters,
       this.operationSpec.parameters,
       // TODO: fix pointer
-    ).map((paramOrRef) => new FieldModel(this.parser, paramOrRef, this.pointer, this.options));
+    ).map(paramOrRef => new FieldModel(this.parser, paramOrRef, this.pointer, this.options));
 
     if (this.options.sortPropsAlphabetically) {
       return sortByField(_parameters, 'name');
@@ -234,7 +239,7 @@ export class OperationModel implements IMenuItem {
   get responses() {
     let hasSuccessResponses = false;
     return Object.keys(this.operationSpec.responses || [])
-      .filter((code) => {
+      .filter(code => {
         if (code === 'default') {
           return true;
         }
@@ -245,7 +250,7 @@ export class OperationModel implements IMenuItem {
 
         return isStatusCode(code);
       }) // filter out other props (e.g. x-props)
-      .map((code) => {
+      .map(code => {
         return new ResponseModel({
           parser: this.parser,
           code,
@@ -259,7 +264,7 @@ export class OperationModel implements IMenuItem {
 
   @memoize
   get callbacks() {
-    return Object.keys(this.operationSpec.callbacks || []).map((callbackEventName) => {
+    return Object.keys(this.operationSpec.callbacks || []).map(callbackEventName => {
       return new CallbackModel(
         this.parser,
         callbackEventName,